Add a new header file for kernel types
